refactor: move common layout to _app and ExampleLayout
