+++ site/pages/_document.tsx
@@ -5,8 +5,6 @@ export default function Document() {
   return (
     <Html lang="en">
       <Head>
-        <title>Slate Examples</title>
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
         <link rel="icon" href="/favicon.ico" />
         <link rel="stylesheet" href="/index.css" />
         <link rel="preconnect" href="https://fonts.googleapis.com" />
